requirements.txt
# AgentCraft — Python dependencies
# Install: pip install -r requirements.txt

# ── Agent execution ───────────────────────────────────────────────────────────
aider-install>=0.2.0      # AI coding assistant installer (drives agents)
pyyaml>=6.0               # workspace.yaml and config parsing
httpx>=0.27.0             # Ollama API calls

# ── RAG ───────────────────────────────────────────────────────────────────────
lancedb>=0.6.0            # embedded vector store (no server)
pyarrow>=15.0             # LanceDB dependency

# ── System monitoring ─────────────────────────────────────────────────────────
psutil>=5.9.0             # CPU/RAM/process metrics (optional but recommended)

# ── Comms server ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
aiosqlite>=0.20.0
